refactor: create parser package
